lsfd: (refactor) move the error object related code to a new file
