+++ lsfd-cmd/Makemodule.am
@@ -11,6 +11,7 @@ lsfd_SOURCES = \
 	lsfd-cmd/decode-file-flags.c \
 	lsfd-cmd/decode-file-flags.h \
 	lsfd-cmd/file.c \
+	lsfd-cmd/error.c \
 	lsfd-cmd/cdev.c \
 	lsfd-cmd/bdev.c \
 	lsfd-cmd/sock.c \
